Major Problem due to memmapfile not accepting complex signal. Fixed and back to working on EK80 data...
Former-commit-id: a539875b95e1b31f6a90c91c7cbd50a382b73a31

+++ @ac_data_cl/get_subdatamat.m
@@ -11,7 +11,7 @@ function datamat=get_subdatamat(data,field,idx_r,idx_ping)
 [idx,found]=find_field_idx(data,(deblank(field)));
 
 if found
-    datamat=data.SubData(idx).Memap.Data.(lower(deblank(field)))(idx_r,idx_ping);
+    datamat=double(data.SubData(idx).Memap.Data.(lower(deblank(field)))(idx_r,idx_ping));
 else
     datamat=[];
 end
